Refactorización
- Evitar buscar recetas por nombre con la cadena vacía.
- Eliminado getRatings de Recipe por un problema de dependencias
circulares
- Arreglado getRecipesByNameOrSteps para que no distinga mayúsculas
- Arreglado getRecipesBySimilarRecipe para que no devuelva a sí misma
- Renombrado storaje.js --> storage.js
- Añadido el ID de receta en recipeToDiv
