+++ src/js/model/Recipe.js
@@ -1,5 +1,4 @@
 const Ingredient = require("./Ingredient");
-const { getState } = require("../appState");
 
 /**
  * Creamos la clase receta.
@@ -84,10 +83,6 @@ class Recipe {
 		} else {
 			throw new Error(tipoDatoInvalido);
 		}
-	}
-
-	get ratings() {
-		return getState().ratings.filter(it => it.idRecipe === this.idRecipe);
 	}
 }
